Test : useLayoutEffect를 통한 flickering 수정
